Split esaggs and request handler.
